Fix issue with Validator toBytes() method.

+++ src/Cohensive/Upload/Validator.php
@@ -338,13 +338,15 @@ class Validator
      */
     protected function toBytes($str)
     {
-        $val = trim($str);
+        $val = (int) trim($str);
         $last = strtolower($str[strlen($str)-1]);
+
         switch($last) {
             case 'g': $val *= 1024;
             case 'm': $val *= 1024;
             case 'k': $val *= 1024;
         }
+
         return $val;
     }
 }
